project/requirements.txt
Django
Fabric
django-webpack-loader
djangorestframework
markdown
django-filter
